requirements.txt
numpy>=1.26.0
obspy>=1.4.0
pyrocko>=2023.10.11
pyyaml>=6.0.1
